refactor: remove the utils folder
