+++ ui/src/entities/User/index.ts
@@ -1,5 +1,5 @@
 import client from 'src/lib/client';
-import resolveStorageFile from 'src/utils/resolveStorageFile';
+import resolveStorageFile from 'src/lib/resolveStorageFile';
 import { Employee } from '../Employee';
 import { Organization } from '../Organizations';
 import { Project } from '../Project';
