+++ ui/src/lib/filterOptionalValues.ts
