+++ ui/src/lib/uploadDefaultLogo.ts
@@ -1,5 +1,5 @@
 import createSvgImage from '../lib/createSvgImage';
-import uploadFile from './uploadFile';
+import uploadFile from '../lib/uploadFile';
 
 export default async (name: string) => {
   const svg = createSvgImage(name);
