+++ ui/src/pages/projects/[id]/index.tsx
@@ -12,7 +12,7 @@ import { fetchProject, projectSelector } from 'src/duck/projects';
 import { updateProjectAction } from 'src/duck/projects';
 import { currentUser } from 'src/entities/User/selectors';
 import { useAsyncThunk, useQuery } from 'src/lib/asyncHooks';
-import filterOptionalValues from 'src/utils/filterOptionalValues';
+import filterOptionalValues from 'src/lib/filterOptionalValues';
 import style from '../style.module.scss';
 
 const c = classNames.bind(style);
