+++ ui/src/pages/projects/[id]/teams/add.tsx
@@ -10,7 +10,7 @@ import { RootState } from 'src/duck';
 import { fetchProject, projectSelector } from 'src/duck/projects';
 import { createTeamAction } from 'src/duck/teams';
 import { useAsyncThunk, useQuery } from 'src/lib/asyncHooks';
-import filterOptionalValues from 'src/utils/filterOptionalValues';
+import filterOptionalValues from 'src/lib/filterOptionalValues';
 import style from './style.module.scss';
 
 const c = classNames.bind(style);
