+++ ui/src/pages/projects/add.tsx
@@ -9,7 +9,7 @@ import { fetchOrganizations, orgSelector } from 'src/duck/organizations';
 import { createProjectAction } from 'src/duck/projects';
 import { currentUser } from 'src/entities/User/selectors';
 import { useAsyncThunk, useQuery } from 'src/lib/asyncHooks';
-import filterOptionalValues from 'src/utils/filterOptionalValues';
+import filterOptionalValues from 'src/lib/filterOptionalValues';
 import style from './style.module.scss';
 
 const c = classNames.bind(style);
